Calculation flow purchase order

+++ Indotalent/DTOs/GoodsReceiveItemChildDto.cs
@@ -7,7 +7,7 @@
         public DateTime? CreatedAtUtc { get; set; }
         public int? WarehouseId { get; set; }
         public int? ProductId { get; set; }
-        public double RequestedMovement { get; set; }
-        public double? Movement { get; set; }
+        public decimal RequestedMovement { get; set; }
+        public decimal? Movement { get; set; }
     }
 }
