+++ Indotalent/DTOs/SalesOrderItemDto.cs
@@ -14,8 +14,8 @@ namespace Indotalent.DTOs
         public int ProductId { get; set; }
         public string? Product { get; set; }
         public string? Summary { get; set; }
-        public double? UnitPrice { get; set; }
-        public double? Quantity { get; set; }
-        public double? Total { get; set; }
+        public decimal? UnitPrice { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? Total { get; set; }
     }
 }
